+++ examples/chatbot/llama_bot_ui.py
@@ -4,7 +4,6 @@ import lightning.app.frontend as frontend
 from chatserver.ui import ui_render_fn
 from llama_inference.serve import PromptRequest, Response, ServeLLaMA
 
-
 checkpoint_path = "weights/state_dict.pth"
 tokenizer_path = "weights/tokenizer.model"
 
@@ -17,7 +16,7 @@ class ChatBotApp(L.LightningFlow):
             output_type=Response,
             checkpoint_path=checkpoint_path,
             tokenizer_path=tokenizer_path,
-        )  
+        )
         self.llm_url = ""
 
     def run(self):
